pyproject.toml
[project]
name = "negbiodb"
version = "0.1.0"
description = "Negative Results Database for Drug-Target Interactions, Clinical Trial Failures, Protein-Protein Interactions, Gene Essentiality, and Variant Pathogenicity"
license = {text = "CC-BY-SA-4.0"}
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.0",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "tqdm>=4.60",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "requests>=2.28",
    "rdkit>=2024.3",
    "chembl-downloader>=0.4",
    "mlcroissant>=1.0",
    "rapidfuzz>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
ml = [
    "torch>=2.2",
    "torch-geometric>=2.5",
    "xgboost>=2.0",
    "pymetis>=2023.1",
    # torch-scatter and torch-sparse require CUDA-specific wheels;
    # install manually on HPC: see slurm/setup_env.sh
]
vp = [
    "torch>=2.2",
    "torch-geometric>=2.4",
    "fair-esm>=2.0",
    "xgboost>=2.0",
]
rltrain = [
    "trl>=1.0.0",
    "peft>=0.14.0",
    "transformers>=4.48.0",
    "accelerate>=1.2.0",
    "datasets>=3.2.0",
    "wandb>=0.19.0",
    "torch>=2.2",
    "vllm>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/negbiodb", "src/negbiodb_ct", "src/negbiodb_ppi", "src/negbiodb_depmap", "src/negbiodb_vp", "src/negbiorl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = ["integration: marks tests requiring network access (deselect with '-m \"not integration\"')"]
